requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
ibind==0.1.19
requests==2.32.5
python-dotenv==1.1.1
pydantic==2.11.7
websocket-client==1.8.0
pycryptodome==3.23.0
